+++ src/main/java/net/frozenblock/lib/config/api/entry/TypedEntry.java
@@ -19,9 +19,24 @@
 package net.frozenblock.lib.config.api.entry;
 
 // java records are meant to be immutable but i dont care
-public record TypedEntry<T>(TypedEntryType<T> type, T value) {
+public class TypedEntry<T>{
+	public final TypedEntryType<T> type;
+	private T value;
 
-    public void setValue(T value) {
+    public TypedEntry(TypedEntryType<T> type, T value) {
+        this.type = type;
         this.value = value;
     }
+
+	public T value() {
+		return this.value;
+	}
+
+	public void setValue(T value) {
+		this.value = value;
+	}
+
+	public TypedEntryType<T> type() {
+		return this.type;
+	}
 }