+++ src/main/kotlin/net/frozenblock/lib/config/api/client/gui/TypedEntryUtils.kt
@@ -57,7 +57,7 @@ fun <T> typedEntryList(
         // Name
         title,
         // Value
-        typedEntry.value,
+        typedEntry.value(),
         // Expanded By Default
         expandedByDefault,
         // Tooltip Supplier