+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Put changelog here:
 -----------------
 - Fixed a crash that would occur on certain versions.
 - Fixed the Experimental Settings config to actually disable the menu on 1.20.5.
+- Added the `FrozenSpawnPlacementTypes` class, containing `ON_GROUND_OR_ON_LAVA_SURFACE` currently.